Formal/casual judgement function implemented
Overview:
We prioritized addressing two categories that are difficult for users to recognize: Formal/Casual and Fashion Genre. We defined the two as opposing elements, meaning "casual" is defined as "NO formal." In a 726-dimensional space, the distribution range for the formal vector is theoretically (-1, 1), but its actual distribution is much narrower. To address this, the creators experimentally set a MAX (most formal) and a MIN (most casual) value. Once this range is established, normalization becomes possible, making the attribute easier to handle. Notably, the term "casual" has a broader definition than "formal," which leads to greater variability in similarity scores for formal attributes.

Challenge 1:
The formal axis can be strengthened by using the average of multiple text vectors. Applying this method could help avoid issues like the overly broad scope observed with "casual," which might be crucial for implementing detailed style and genre classifications.

Challenge 2:
Accuracy and user experience can also be enhanced through additional training (fine-tuning) or by adopting or combining different models.